requirements.txt
keras==2.3.1
opencv-python
scipy==1.1.0
flask==1.1.1
gunicorn
tensorflow==1.15.2
pillow==7.1.1